Library API

Library API

Spotify.library_get_saved_shows — Method

Get a User's Saved Shows

Summary: Get a list of the songs saved in the current Spotify user's 'Your Music' library.

limit Optional: The maximum number of objects to return. Default 20. Minimum 1. Maximum 50.

offset Optional: The index of the first object to return. Default 0.

Reference

Spotify.library_check_saved_albums — Method

Check User's Saved Albums

Summary: Check if one or more albums is already saved in the current Spotify user's 'Your Music' library.

album_ids Required: A comma separated list of the Spotify IDs for the albums. Max 50.

Required

Spotify.library_check_saved_shows — Method

Check User's Saved Shows

Summary: Check if one or more shows is already saved in the current Spotify user's library.

show_ids Required: A comma separated list of the Spotify IDs for the shows. Maximum 50

Reference

Spotify.library_check_saved_tracks — Method

Check User's Saved Tracks

Summary: Check if one or more tracks is already saved in the current Spotify user's 'Your Music' library.

track_ids Required: A comma-separated list of the Spotify IDs for the tracks. Maximum 50.

Reference

Spotify.library_get_saved_albums — Function

Get a User's Saved Albums

Summary: Get a list of the albums saved in the current Spotify user's 'Your Music' library.

limit Optional: The maximum number of objects to return. Default 20. Minimum 1. Maximum 50.

offset Optional: The index of the first object to return. Default 0.

market Optional: An ISO 3166-1 alpha-2 country code.

Reference

Spotify.library_get_saved_shows — Function

Get a User's Saved Shows

Summary: Get a list of the songs saved in the current Spotify user's 'Your Music' library.

limit Optional: The maximum number of objects to return. Default 20. Minimum 1. Maximum 50.

offset Optional: The index of the first object to return. Default 0.

Reference

Spotify.library_get_saved_tracks — Function

Get a User's Saved Tracks

Summary: Get a list of the songs saved in the current Spotify user's 'Your Music' library.

limit Optional: The maximum number of objects to return. Default 20. Minimum 1. Maximum 50.

offset Optional: The index of the first object to return. Default 0.

market Optional: An ISO 3166-1 alpha-2 country code.

Reference

Spotify.library_remove_albums — Method

Remove Albums for Current User

Summary: Remove one or more albums for the current user's 'Your Music' library.

album_ids Required: A comma-separated list of the Spotify IDs. Maximum 50.

Reference

Spotify.library_remove_shows — Method

Remove Shows for Current User

Summary: Remove one or more shows for the current user's library.

show_ids Required: A comma-separated list of the Spotify IDs. Maximum 50.

Reference

Spotify.library_remove_tracks — Method

Remove Tracks for Current User

Summary: Remove one or more tracks for the current user's 'Your Music' library.

track_ids Required: A comma-separated list of the Spotify IDs. Maximum 50.

Reference

Spotify.library_save_album — Method

Save Albums for Current User

** Summary**: Save one or more albums to the current user's 'Your Music' library.

album_ids Required: A comma-separated list of Spotify IDs. Maximum 50.

Reference

Spotify.library_save_show — Method

Save Shows for Current User

** Summary**: Save one or more shows to the current user's library.

shows_ids Required: A comma-separated list of Spotify IDs. Maximum 50.

Reference

Spotify.library_save_track — Method

Save Tracks for Current User

** Summary**: Save one or more tracks to the current user's 'Your Music' library.

track_ids Required: A comma-separated list of Spotify IDs. Maximum 50.

Reference