Reference

Spotify.@_ie — Macro

import export shorthand

Spotify._get_ini_fnam — Method

Get an existing, readable ini file name, create it if necessary

Spotify.album_get — Function

Get an Album(s)

Summary: Get Spotify catalog information for an album(s)

album_id Required: The Spotify ID for the album. Up to 20 albums can be passed by comma separating the ID's

market Optional: An ISO 3166-1 alpha-2 country code. Default "US"

Reference

Spotify.album_get_tracks — Function

Get an Album's Tracks

Summary: Get Spotify catalog information about an album's tracks. Optional parameters can be used to limit the number of tracks returned

album_id Required: The Spotify ID for the album. Up to 20 albums can be passed by limit Optional: The maximum number of tracks to return. Default 20

offset Optional: The index of the first track to return. Default 0

market Optional: An ISO 3166-1 alpha-2 country code. Default "US"

Reference

Spotify.artist_get — Method

Get an Artist

Summary: Get Spotify catalog information for a single artist identified by their unique Spotify ID.

artist_id Required: The Spotify artist ID. Up to 50 artist ID's can be passed by comma delimiting the ID's

Reference

Spotify.artist_get_albums — Function

Get an Artists Albums

Summary: Get Spotify catalog information about an artist's albums.

artist_id Required: The Spotify artist ID. Up to 50 artist ID's can be passed by comma delimiting the ID's

include_groups Optional: A comma-separated list of keywords that will be used to filter the response. If not supoplied, all album types will be returned. Valid values:

album

single

compilation

appears_on

country Optional: An ISO 3166-1 alpha-2 country code string. Use this to limit the response to one particular geographical market. Default "US"

limit Optional: The number of album objects to return. Default 20

offset Optional: The index of the first album to return. Default 0

Reference

Spotify.artist_get_related_artists — Method

Get an Artist's Related Artists

Summary: Get spotify catalog information about artists similar to a given artist. Similarity is based on analysis of the Spotify community's listening history.

artist_id Required: The Spotify artist ID.

Reference

Spotify.artist_top_tracks — Function

Get an Artist's Top Tracks

Summary: Get Spotify catalog information about an artist's top tracks by country.

artist_id Required: The Spotify artist ID.

country Optional: An ISO 3166-1 alpha-2 country code string. Use this to limit the response to one particular geographical market. Default "US"

Reference

Spotify.authorize — Method

authorize()

Get and store client credentials. Any other credentials will be dropped.

Spotify.category_get — Function

" Get a single category used to tag items in Spotify (on, for example, the Spotify player’s “Browse” tab).

country: (Optional) A country: ISO 3166-1 alpha-2 country code. Provide this parameter to ensure that the category exists for a particular country.

locale: (Optional) The desired language, consisting of an ISO 639-1 language code and an ISO 3166-1 alpha-2 country code, joined by an underscore. For example: es_MX, meaning "Spanish (Mexico)". Provide this parameter if you want the category strings returned in a particular language. Note that, if locale is not supplied, or if the specified language is not available, the category strings returned will be in the Spotify default language (American English).

Spotify.category_get_featured_playlist — Function

Get a list of Spotify featured playlists (shown, for example, on a Spotify player’s ‘Browse’ tab).

country: (Optional) A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

locale: (Optional) The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset: (Optional) The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

timestamp: (Optional) A timestamp in ISO 8601 format: yyyy-MM-ddTHH:mm:ss. Use this parameter to specify the user’s local time to get results tailored for that specific date and time in the day. If not provided, the response defaults to the current UTC time. Example: “2014-10-23T09:00:00” for a user whose local time is 9AM. If there were no featured playlists (or there is no data) at the specified time, the response will revert to the current UTC time.

Spotify.category_get_new_releases — Function

Get a list of new album releases featured in Spotify (shown, for example, on a Spotify player’s “Browse” tab).

country: (Optional) A country: an ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. If omitted, the returned items will be relevant to all countries.

locale: (Optional) The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset: (Optional) Optional. The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

Spotify.category_get_playlist — Function

Get a list of Spotify playlists tagged with a particular category.

country: (Optional) A country: an ISO 3166-1 alpha-2 country code.

limit: (Optional) The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset: (Optional) The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

Spotify.category_get_several — Function

Get a list of categories used to tag items in Spotify (on, for example, the Spotify player’s “Browse” tab).

country: (Optional) A country: an ISO 3166-1 alpha-2 country code.

limit: (Optional) The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50.

offset: (Optional) The index of the first item to return. Default: 0 (the first object). Use with limit to get the next set of items.

Spotify.episodes_get — Function

Get an Episode

Summary: Get Spotify catalog information for a single peisode identified by it's unique Spotify ID.

episode_id Required: The Spotify ID for the episode_id. Up to 50 episodes can be passed, comma delimited.

market Optional: An ISO 3166-1 alpha 2 country code.

Reference

Spotify.follow_artists — Function

follow_artists(item_type::String="artist", limit::Int64=20)

Summary: Get the current user's followed artists.

Arguments

item_type::String Required: The ID type. Currently only artist is supported. Default artist.

limit::Int64 Optional: The maximum number of items to return. Default 20, Minimum 1, Maximum 50.

Example

julia> Spotify.follow_artists()[1]["artists"]
[ Info: We try the request without checking if current grant includes scope user-follow-modify.
JSON3.Object{Base.CodeUnits{UInt8, String}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}} with 6 entries:
  :items   => JSON3.Object[{…
  :next    => nothing
  :total   => 2
  :cursors => {…
  :limit   => 20
  :href    => "https://api.spotify.com/v1/me/following?type=artist&limit=20"

Spotify.follow_artists_users — Method

Follow Artists or Users

Summary: Add the current user as a follower of one or more artists or other Spotify users.

type Required: The ID type: either artist or user.

ids Required: A comma-separated list of the artists or users Spotify IDs. Maximum 50.

Reference

Spotify.follow_check — Method

follow_check(item_type, ids)

Summary: Check to see if the current user is following one or more artists or other Spotify users.

Arguments

item_type::String Required: The ID type, either artist or user.

ids::String Required: A comma separated list of the artist or user Spotify IDs to check. Maximum 50.

Example

julia> Spotify.follow_check("artist", "7fxBPUc2bTUgl7GLuqjajk")[1]
[ Info: We try the request without checking if current grant includes scope user-follow-read.
1-element JSON3.Array{Bool, Base.CodeUnits{UInt8, String}, Vector{UInt64}}:
 1

Spotify.follow_check_playlist — Method

follow_check_playlist(playlist_id::String, user_id::String)

Summary: Check to see if one or more Spotify users are following a specified playlist_id.

Arguments

playlist_id::String Required: The Spotify ID of the playlist.

user_id::String Required: A comma separated list of the user Spotify IDS to check. Maximum 5.

Example

julia> Spotify.follow_check_playlist("37i9dQZF1DZ06evO3Khq6I", user_id)[1]
1-element JSON3.Array{Bool, Base.CodeUnits{UInt8, String}, Vector{UInt64}}:
 0

Spotify.follow_playlist — Method

Follow a Playlist

Summary: Add the currend user as a follower of a playlist.

playlist_id Required: The Spotify ID of the playlist. Any playlist can be followed regardless of it's private/public status, as long as the ID is known.

Reference

Spotify.fwhich — Method

Get application path for developer applications

Spotify.genres_get — Method

genres_get()

Summary: Retrieve a list of available genres seed parameter values for recommendations.

Example

julia> Spotify.genres_get()[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 1 entry:
  :genres => ["acoustic", "afrobeat", "alt-rock", "alternative", "ambient", "anime", "black-metal",…

Spotify.get_id_secret_redirect — Method

Get id and secret as 32-byte string, no encryption

Spotify.get_implicit_grant — Method

If the user gives us access, that will be in the future. Is it worth waiting for? If so, call wait_for_ig_access()

Spotify.launch_command — Method

Constructs launch command

Spotify.markets_get — Method

markets_get()

Summary: Get the list of markets where Spotify is available.

Example

julia> Spotify.markets_get()[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 1 entry:
  :markets => ["AD", "AE", "AG", "AL", "AM", "AO", "AR", "AT", "AU", "AZ"  …

Spotify.max_acousticness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_danceability — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_duration_ms — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_energy — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_instrumentalness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_key — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_liveness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_loudness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_mode — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_popularity — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_speechiness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_tempo — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_time_signature — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_valence — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.min_acousticness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_danceability — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_duration_ms — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_energy — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_instrumentalness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_key — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_liveness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_loudness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_mode — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_popularity — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_speechiness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_tempo — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_time_signature — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_valence — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.open_a_browser — Method

Try to open one browser from BROWSERS. In some cases we expect an immediate indication of failure, for example when the corresponding browser is not found on the system. In other cases, we will just wait in vain. In those cases, call this function again after a reasonable timeout. The function remembers which browsers were tried before.

Spotify.playlist_get — Method

playlist_get(playlist_id::String; additional_types::String="track", fields::String="",
market::String="US")

Summary: Get details about a playlist owned by a Spotify user.

Arguments

playlist_id::String : Alphanumeric ID of the playlist

Optional keywords

additional_types::String : "track" (default) or "episode"

fields::String : Filters for the query, a comma-separated list of the fields to return. For example, to get just the added date and user ID of the adder, fields = "items(addedat,addedby.id)". Default is set to "", which means all fields are returned.

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

Example

julia> Spotify.playlist_get("37i9dQZF1E4vUblDJbCkV3")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 15 entries:
  :collaborative => false
  :description   => "With Roo Panes, Hiss Golden Messenger, Nathaniel Rateliff and more"
  :external_urls => {…
  :followers     => {…
  :href          => "https://api.spotify.com/v1/playlists/37i9dQZF1E4vUblDJbCkV3?additional_types=track"
  :id            => "37i9dQZF1E4vUblDJbCkV3"
  :images        => JSON3.Object[{…
  :name          => "Lord Huron Radio"
  :owner         => {…
  :primary_color => nothing
  :public        => false
  :snapshot_id   => "MTYzNTg2NzMxNCwwMDAwMDAwMGU3MTgwNDkzOWE5NTQ2NGM1NmYzNTYyZDhhZTc1ZGNh"
  :tracks        => {…
  :type          => "playlist"
  :uri           => "spotify:playlist:37i9dQZF1E4vUblDJbCkV3"

Spotify.playlist_get_category — Method

playlist_get_category(category_id::String; country::String="US", limit::Int64=20, offset::Int64=0)

Summary: Get a list of Spotify playlists tagged with a particular category.

Arguments

category_id::String : The unique string identifying the Spotify category, e.g. "dinner", "party" etc.

Optional keywords

country::String : An ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. Default is set to "US".

limit::Int64 : Maximum number of items to return, default is set to 20

offset::Int64 : Index of the first item to return, default is set to 0

Example

julia> Spotify.playlist_get_category("party")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 1 entry:
  :playlists => {…

Spotify.playlist_get_cover_image — Method

playlist_get_cover_image(playlist_id::String)

Summary: Get the current image associated with a specific playlist.

Arguments

playlist_id::String : Alphanumeric ID of the playlist

Example

julia> Spotify.playlist_get_cover_image("37i9dQZF1E4vUblDJbCkV3")[1]
1-element JSON3.Array{JSON3.Object, Base.CodeUnits{UInt8, String}, Vector{UInt64}}:
 {
   "height": nothing,
      "url": "https://seeded-session-images.scdn.co/v1/img/artist/6ltzsmQQbmdoHHbLZ4ZN25/en",
    "width": nothing
}

Spotify.playlist_get_current_user — Method

playlist_get_current_user(limit::Int64=20, offset::Int64=0)

Summary: Get a list of the playlists owned or followed by the current Spotify user.

Optional keywords

limit::Int64 : Maximum number of items to return, default is set to 20

offset::Int64 : Index of the first item to return, default is set to 0

Example

julia> Spotify.playlist_get_current_user()[1]
[ Info: We try the request without checking if current grant includes scope playlist-read-private.
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 7 entries:
  :href     => "https://api.spotify.com/v1/users/your_user_id/playlists?offset=0&limit=20"
  :items    => JSON3.Object[{…
  :limit    => 20
  :next     => nothing
  :offset   => 0
  :previous => nothing
  :total    => 2

Spotify.playlist_get_featured — Method

playlist_get_featured(;country::String="US", limit::Int64=20, locale::String="en_US", 
                           offset::Int64=0, timestamp::String="2021-11-09T17:28:14.297")

Summary: Get a list of Spotify featured playlists (shown, for example, on a Spotify player's 'Browse' tab).

Optional keywords

country::String : An ISO 3166-1 alpha-2 country code. Provide this parameter if you want the list of returned items to be relevant to a particular country. Default is set to "US".

limit::Int64 : Maximum number of items to return, default is set to 20

locale::String : The desired language, consisting of a lowercase ISO 639-1 language code and an uppercase ISO 3166-1 alpha-2 country code, joined by an underscore. For example: esMX, meaning "Spanish (Mexico)". Provide this parameter if you want the results returned in a particular language (where available). Default is set to "enUS".

offset::Int64 : Index of the first item to return, default is set to 0

timestamp::String : A timestamp in ISO 8601 format: yyyy-MM-ddTHH:mm:ss. Use this parameter to specify the user's local time to get results tailored for that specific date and time in the day. Default is set to user's current time.

Example

julia> Spotify.playlist_get_featured(locale="en_UK")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 2 entries:
  :message   => "Tuesday jams"
  :playlists => {…

Spotify.playlist_get_tracks — Method

playlist_get_tracks(playlist_id::String; additional_types="track", limit=50, 
offset=0, market="US")

Summary: Get details about the items of a playlist.

Arguments

playlist_id::String : Alphanumeric ID of the playlist

Optional keywords

additional_types::String : "track" (default) or "episode"

fields::String : Filters for the query, a comma-separated list of the fields to return. For example, to get just the added date and user ID of the adder, fields = "items(addedat,addedby.id)". Default is set to "", which means all fields are returned.

limit::Int64 : Maximum number of items to return, default is set to 20

offset::Int64 : Index of the first item to return, default is set to 0

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

Example

julia> Spotify.playlist_get_tracks("37i9dQZF1E4vUblDJbCkV3")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 7 entries:
  :href     => "https://api.spotify.com/v1/playlists/37i9dQZF1E4vUblDJbCkV3/tracks?offset=0&limit=20&market=US&additional_types=…
  :items    => JSON3.Object[{…
  :limit    => 20
  :next     => "https://api.spotify.com/v1/playlists/37i9dQZF1E4vUblDJbCkV3/tracks?offset=20&limit=20&market=US&additional_types…
  :offset   => 0
  :previous => nothing
  :total    => 50

Spotify.playlist_get_user — Method

playlist_get_user(user_id::String; limit::Int64=20, offset::Int64=0)

Summary: Get a list of the playlists owned or followed by a Spotify user.

Arguments

user_id::String : Alphanumeric ID of the user or name (e.g. "smedjan")

Optional keywords

limit::Int64 : Maximum number of items to return, default is set to 20

offset::Int64 : Index of the first item to return, default is set to 0

Example

julia> Spotify.playlist_get_user("smedjan")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 7 entries:
  :href     => "https://api.spotify.com/v1/users/smedjan/playlists?offset=0&limit=20"
  :items    => JSON3.Object[{…
  :limit    => 20
  :next     => "https://api.spotify.com/v1/users/smedjan/playlists?offset=20&limit=20"
  :offset   => 0
  :previous => nothing
  :total    => 98

Spotify.recommendations_get — Function

Get Recommendations Based on Seeds

seeds: (Required) A dictionary containing keys(seedgenres, seedartists, seed_tracks) and values for each key being seeds delimeted by a comma up to 5 seeds for each category

Example

Dict("seedartists" => "s33dart1st,s33edart!st2", "seedgenres" => "g3nre1,genr32", "seed_tracks" => "trackid1,trackid2")

trackattributes: (Optional) a dictionary containing key values for different tunable track trackattributes

See https://developer.spotify.com/documentation/web-api/reference/browse/get-recommendations/

Spotify.search_get — Method

search_get(;q::String="Coldplay", item_type::String="track,artist", include_external::String="",
                limit::Int64=20, market::String="US", offset::Int64=0)

Summary: Get Spotify catalog information about albums, artists, playlists, tracks, shows or episodes that match a keyword string.

Optional keywords

q::String : Search query, e.g. "Coldplay" which is also set as the default

item_type::String : A comma-separated list of item types to search across. Search results include hits from all the specified item types. For example, item_type = "album,tarck" returns both albums and tracks with the search query included in their name.

include_external::String : If include_external="audio" is specified then the response will include any relevant audio content that is hosted externally.

limit::Int64 : Maximum number of items to return, default is set to 20

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

offset::Int64 : Index of the first item to return, default is set to 0

Example

julia> Spotify.search_get(;q = "Greenday", item_type = "album")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 1 entry:
  :albums => {…

Spotify.show_get — Function

Get a Show

Summary: Get a Spotify catalog information for a single show identified by it's unique Spotify ID.

show_id Required: The Spotify ID for the show. Up to 50 shows can be passed seperated with a comma. No whitespace.

market Optional: An ISO 3166-1 alpha-2 country code. If a country code is specified, only shows and episodes that are available in that market will be returned.

Reference

Spotify.show_get_episodes — Function

Get a Show's Episodes

Summary: Get Spotify catalog information about a show’s episodes. Optional parameters can be used to limit the number of episodes returned.

show_id Required: The Spotify ID for the show.

market Optional: Optional. An ISO 3166-1 alpha-2 country code. If a country code is specified, only shows and episodes that are available in that market will be returned.

limit Optional: The maximum number of episodes to return. Default = 20. Minimum 1. Maximum 50 offset Optional: The index of the first episode to return. Default: 0 (the first object). Use the limit to get the next set of episodes.

Reference

Spotify.spotify_request — Function

spotify_request(url_ext::String, method::String= "GET"; scope = "client-credentials")
 -> (r::JSON3.Object, retry_in_seconds::Int)

Access the Spotify Web API. Error results return an empty Object. Errors are written to 'stderr', expect for 'API rate limit exceeded', as the output would typically occur in the middle of recursive calls.

Spotify.strip_embed_code — Method

strip_embed_code(sdvs<substring>)
-> Spid(<substring>)

Get the interesting part for pasting:

Spotify app -> Right click -> Share -> Copy embed code to clipboard

Spotify.target_acousticness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_danceability — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_duration_ms — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_energy — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_instrumentalness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_key — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_liveness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_loudness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_mode — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_popularity — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_speechiness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_tempo — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_time_signature — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_valence — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.top_artists — Method

top_artists(;offset=0, limit=20, time_range="medium_term")

Summary: Get the current user's top artists based on calculated affinity.

Optional keywords

offset Optional: The index of the first tracks to return. Default 0.

limit Optional: The number of tracks to return. Default 20.

time_range Optional: Over what time frame the affinities are computed. Valid Options: * long_term : Calculated from several years of data including all new data as it becomes available * medium_term : Approx. last 6 months * short_term : Approx. last 4 weeks

Spotify.top_tracks — Method

top_tracks(;offset=0, limit=20, time_range="medium_term")

Summary: Get the current user's top tracks based on calculated affinity.

Optional keywords

offset Optional: The index of the first tracks to return. Default 0.

limit Optional: The number of tracks to return. Default 20.

time_range Optional: Over what time frame the affinities are computed. Valid Options: * long_term : Calculated from several years of data including all new data as it becomes available * medium_term : Approx. last 6 months * short_term : Approx. last 4 weeks

Spotify.tracks_get — Function

Get a Track

Summary: Get a spotify catalog information for a single track identified by it's unique Spotify ID

track_id Required: The Spotify ID for the track. Up to 50 tracks can be passed seperated with a comma. No whitespace.

market Optional: An ISO 3166=1 alpha-2 country code

Reference

Spotify.tracks_get_audio_analysis — Method

Get Audio Analysis for a Track

Summary: Get a detailed audio analysis for a single track identified by it's unique Spotify ID.

track_id Required: The Spotify ID for the track.

Reference

Spotify.tracks_get_audio_features — Method

Get Audio Features for a Track

Summary: Get audio feature information for a single track identified by it's unique Spotify ID

track_id Required: The Spotify ID for the track. Up to 50 tracks can be passed seperated with a comma. No whitespace.

Reference

Spotify.unfollow_artists_users — Method

Unfollow Artists or Users

Summary: Remove the current user as a follower of one or more artists or other Spotify users.

type Required: The ID type: either artist or user.

ids Required: A comma-separated list of the artists or users Spotify IDs. Maximum 50.

Reference

Spotify.unfollow_playlist — Method

Unfollow a Playlist

Summary: Remove the current user as a follower of a playlist.

playlist_id Required: The Spotify ID of the playlist. Any playlist can be followed regardless of it's private/public status, as long as the ID is known.

Reference

Spotify.user_get_current_profile — Method

user_get_current_profile()

**Summary**: Get detailed profile information about the current user 
            (including the current user's username).

Example

julia> Spotify.user_get_current_profile()[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 12 entries:
  :country          => "NL"
  :display_name     => "Itachi"
  :email            => "your_id@gmail.com"
  :explicit_content => {…
  :external_urls    => {…
  :followers        => {…
  :href             => "https://api.spotify.com/v1/users/your_user_id"
  :id               => "your_user_id"
  :images           => Union{}[]
  :product          => "premium"
  :type             => "user"
  :uri              => "spotify:user:your_user_id"

Spotify.user_get_profile — Method

user_get_profile(user_id::String)

Summary: Get public profile information about a Spotify user.

Arguments

user_id::String : Alphanumeric ID of the user or name (e.g. "smedjan")

Example

julia> Spotify.user_get_profile("smedjan")[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 8 entries:
  :display_name  => "smedjan"
  :external_urls => {…
  :followers     => {…
  :href          => "https://api.spotify.com/v1/users/smedjan"
  :id            => "smedjan"
  :images        => Union{}[]
  :type          => "user"
  :uri           => "spotify:user:smedjan"

Spotify.Countbrowser — Type

A stateful browser counter.

Spotify.SpCategoryId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpPlaylistId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUri — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUrl — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUserId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.AUTH_URL — Constant

For the client credentials flow

Spotify.BROWSERS — Constant

A list of potentially available browsers, to be tried in succession if present

Spotify.COUNTBROWSER — Constant

For next value: COUNTBROWSER(). For current value: COUNTBROWSER.value

Spotify.DEFAULT_IMPLICIT_GRANT — Constant

These permissions are not requested until the current scope is insufficient, or the user calls 'getimplicitgrant()'

Default requested permissions are 'client-credentials'.

Spotify.OAUTH_AUTHORIZE_URL — Constant

For the 'Implicit grant flow