Reference

Spotify.@_ie — Macro

import export shorthand

Spotify._get_ini_fnam — Method

Get an existing, readable ini file name, create it if necessary

Spotify.authorize — Method

authorize()

Get and store client credentials. Any other credentials will be dropped.

Spotify.get_id_secret_redirect — Method

Get id and secret as 32-byte string, no encryption

Spotify.get_implicit_grant — Method

If the user gives us access, that will be in the future. Is it worth waiting for? If so, call wait_for_ig_access()

Spotify.max_acousticness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_danceability — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_duration_ms — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_energy — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_instrumentalness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_key — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_liveness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_loudness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_mode — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_popularity — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_speechiness — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_tempo — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_time_signature — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.max_valence — Function

For each tunable track attribute, a hard ceiling on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, max_instrumentalness=0.35 would filter out most tracks that are likely to be instrumental.

Spotify.min_acousticness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_danceability — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_duration_ms — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_energy — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_instrumentalness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_key — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_liveness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_loudness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_mode — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_popularity — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_speechiness — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_tempo — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_time_signature — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.min_valence — Function

For each tunable track attribute, a hard floor on the selected track attribute’s value can be provided. See tunable track attributes below for the list of available options. For example, min_tempo=140 would restrict results to only those tracks with a tempo of greater than 140 beats per minute.

Spotify.player_get_current_track — Method

player_get_current_track(;additional_types::String="track", market::String="US")

Summary: Get the object currently being played on the user's Spotify account.

Optional keywords

additional_types::String : "track" (default) or "episode"

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

Example

julia> Spotify.player_get_current_track()[1]
[ Info: We try the request without checking if current grant includes scope user-read-playback-state.
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 7 entries:
  :timestamp              => 1636491068506
  :context                => {…
  :progress_ms            => 5265
  :item                   => {…
  :currently_playing_type => "track"
  :actions                => {…
  :is_playing             => true

Spotify.player_get_devices — Method

player_get_devices()

Summary: Get information about a user’s available devices.

Example

julia> Spotify.player_get_devices()[1]["devices"]
2-element JSON3.Array{JSON3.Object, Base.CodeUnits{UInt8, String}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}}:
{
                   "id": "your_device_id",
            "is_active": false,
   "is_private_session": false,
        "is_restricted": false,
                 "name": "Web Player (Chrome)",
                 "type": "Computer",
       "volume_percent": 100
}

Spotify.player_get_state — Method

player_get_state(;additional_types::String="track", market::String="US")

Summary: Get information about the user’s current playback state, including track or episode, progress, and active device.

Optional keywords

additional_types::String : "track" (default) or "episode"

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

Example

julia> Spotify.player_get_state()[1]
[ Info: We try the request without checking if current grant includes scope user-read-playback-state.
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 10 entries:
  :device                 => {…
  :shuffle_state          => false
  :repeat_state           => "off"
  :timestamp              => 1636493367689
  :context                => {…
  :progress_ms            => 66454
  :item                   => {…
  :currently_playing_type => "track"
  :actions                => {…
  :is_playing             => true

Spotify.recommendations_get — Method

recommendations_get(seeds; track_attributes, limit::Int64 = 50, market::String = "US")

Summary: Get Recommendations Based on Seeds

Arguments

seeds : A dictionary containing keys(seedgenres, seedartists, seedtracks) and values for each key being seeds delimited by a comma up to 5 seeds for each category. For example: Dict("seedartists" => "s33dart1st,s33edart!st2", "seedgenres" => "g3nre1,genr32", "seedtracks" => "trackid1,trackid2")

Optional keywords

track_attributes : A dictionary containing key values for different tunable track track_attributes

limit::Int64 : Maximum number of items to return, default is set to 20

market::String : An ISO 3166-1 alpha-2 country code. If a country code is specified, only episodes that are available in that market will be returned. Default is set to "US".

Reference

(https://developer.spotify.com/documentation/web-api/reference/#/operations/get-recommendations)

Example

julia> seeds = Dict("seed_artists" => "0YC192cP3KPCRWx8zr8MfZ")
Dict{String, String} with 1 entry:
  "seed_artists" => "0YC192cP3KPCRWx8zr8MfZ"

julia> track_attributes = Dict("max_danceability" => "0.80")
Dict{String, String} with 1 entry:
  "max_danceability" => "0.80"

julia> Spotify.recommendations_get(seeds, track_attributes = track_attributes)[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, Vector{UInt64}} with 2 entries:
  :tracks => JSON3.Object[{…
  :seeds  => JSON3.Object[{…

Spotify.spotify_request — Function

spotify_request(url_ext::String, method::String= "GET"; scope = "client-credentials")
 -> (r::JSON3.Object, retry_in_seconds::Int)

Access the Spotify Web API. Error results return an empty Object. Errors are written to 'stderr', expect for 'API rate limit exceeded', as the output would typically occur in the middle of recursive calls.

Spotify.strip_embed_code — Method

strip_embed_code(sdvs<substring>)
-> Spid(<substring>)

Get the interesting part for pasting:

Spotify app -> Right click -> Share -> Copy embed code to clipboard

Spotify.target_acousticness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_danceability — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_duration_ms — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_energy — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_instrumentalness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_key — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_liveness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_loudness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_mode — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_popularity — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_speechiness — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_tempo — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_time_signature — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.target_valence — Function

For each of the tunable track attributes (below) a target value may be provided. Tracks with the attribute values nearest to the target values will be preferred. For example, you might request targetenergy=0.6 and targetdanceability=0.8. All target values will be weighed equally in ranking results.

Spotify.SpAlbumId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpArtistId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpCategoryId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpPlaylistId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpShowId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUri — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUrl — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.SpUserId — Type

All web API arguments are strings, but types SpUri, SpId, CategoryId, SpUserId, SpUrl aids in picking default values. format

PARAMETER	DESCRIPTION	VALUE

SpUri	The resource identifier that you can enter, for example, in the

Spotify Desktop client’s search box to locate an artist,

album, or track. To find a Spotify URI simply right-

click (on Windows) or Ctrl-Click (on a Mac) on	spotify:track:

the artist’s or album’s or track’s name.	6rqhFgbbKwnb9MLmUQDhG6

SpId	The base-62 identifier that you can find at the end of the

Spotify URI (see above) for an artist, track, album,

playlist, etc. Unlike a Spotify URI, a Spotify ID does

not clearly identify the type of resource; that information is

provided elsewhere in the call.	6rqhFgbbKwnb9MLmUQDhG6

SpCategoryId	The unique string identifying the Spotify category.	party

SpUserId	The unique string identifying the Spotify user that you can

find at the end of the Spotify URI for the user. The ID

of the current user can be obtained via the Web API endpoint.	wizzler

SpUrl	An HTML link that opens a track, album, app, playlist or other

Spotify resource in a Spotify client (which client

is determined by the user’s device and	http://open.spotify.com/

account settings at play.spotify.com.	track/6rqhFgbbKwnb9MLmUQDhG6

Spotify.AUTH_URL — Constant

For the client credentials flow

Spotify.BROWSERS — Constant

A list of potentially available browsers, to be tried in succession if present

Spotify.DEFAULT_IMPLICIT_GRANT — Constant

These permissions are not requested until the current scope is insufficient, or the user calls 'getimplicitgrant()'

Default requested permissions are 'client-credentials'.

Spotify.OAUTH_AUTHORIZE_URL — Constant

For the 'Implicit grant flow